Add long

+++ backend/Arbiter/Arbiter.BLL/DataTransferObjects/RequestLoanDto.cs
@@ -3,10 +3,10 @@
 namespace Arbiter.BLL.DataTransferObjects;
 
 public class RequestLoanDto
-{   
+{
     public Guid UserId { get; set; }
     public Guid AccountId { get; set; }
     public Guid TariffId { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public CurrencyType CurrencyType { get; set; }
 }
