+++ backend/Arbiter/Arbiter.DAL/Entities/ChargeLoanTransaction.cs
@@ -9,16 +9,15 @@ public class ChargeLoanTransaction
     public Guid AccountId { get; set; }
     public Guid TariffId { get; set; }
     public Guid LoanId { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public CurrencyType CurrencyType { get; set; }
-    
+
     public TransactionStatus CheckAccountStatus { get; set; } = TransactionStatus.NotStarted;
     public int CheckAccountRetries { get; set; }
-    
-    
+
     public TransactionStatus AccountLoanChargeStatus { get; set; } = TransactionStatus.NotStarted;
     public int AccountLoanChargeRetries { get; set; }
-    
+
     public TransactionStatus LoanChargeStatus { get; set; } = TransactionStatus.NotStarted;
     public int LoanChargeRetries { get; set; }
 }
