+++ backend/Arbiter/Arbiter.DAL/Entities/RequestLoanTransaction.cs
@@ -9,15 +9,15 @@ public class RequestLoanTransaction
     public Guid AccountId { get; set; }
     public Guid TariffId { get; set; }
     public Guid LoanId { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public CurrencyType CurrencyType { get; set; }
 
     public TransactionStatus CheckAccountStatus { get; set; } = TransactionStatus.NotStarted;
     public int CheckAccountRetries { get; set; }
-    
-    public TransactionStatus TakeLoanStatus { get; set; }  = TransactionStatus.NotStarted;
+
+    public TransactionStatus TakeLoanStatus { get; set; } = TransactionStatus.NotStarted;
     public int TakeLoanRetries { get; set; }
-    
-    public TransactionStatus AccountLoanIncomeStatus { get; set; }  = TransactionStatus.NotStarted;
+
+    public TransactionStatus AccountLoanIncomeStatus { get; set; } = TransactionStatus.NotStarted;
     public int AccountLoanIncomeRetries { get; set; }
 }
