+++ backend/Core/Core.BLL/DataTransferObjects/AccountDto.cs
@@ -8,10 +8,9 @@ public class AccountDto
     public Guid Id { get; set; }
     public Guid UserId { get; set; }
     public CurrencyType CurrencyType { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public DateTime? ClosedAt { get; set; }
     public bool IsPriority { get; set; } = false;
-
 }
 
 public static class AccountDtoExtensions
