+++ backend/Core/Core.BLL/DataTransferObjects/DepositDto.cs
@@ -1,6 +1,7 @@
 ﻿namespace Core.BLL.DataTransferObjects;
 
-public class DepositDto {
-    public int Amount { get; set; }
+public class DepositDto
+{
+    public long Amount { get; set; }
     public string? Message { get; set; }
 }
