+++ backend/Loan/Loan.BLL/DataTransferObjects/PaymentDto.cs
@@ -1,12 +1,13 @@
 ﻿namespace Loan.BLL.DataTransferObjects;
 
-public class PaymentDto {
+public class PaymentDto
+{
     public Guid Id { get; set; } = Guid.NewGuid();
     public LoanDto Loan { get; set; }
-    public int AmountForPay { get; set; } // сколько нужно заплатить
-    public int AlreadyPaid { get; set; } // сколько заплатил
+    public long AmountForPay { get; set; } // сколько нужно заплатить
+    public long AlreadyPaid { get; set; } // сколько заплатил
     public DateTime? PaidAt { get; set; } // во сколько заплатил
     public bool IsActual { get; set; } // текущая заявка
-    public int PenaltyFee { get; set; }
+    public long PenaltyFee { get; set; }
     public DateTime CreatedAt { get; set; }
 }
